% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfolds.R
\name{bfolds}
\alias{bfolds}
\title{Balanced folds for cross-validation for classification}
\usage{
bfolds(y, nfolds = min(min(table(y)), 10))
}
\arguments{
\item{y}{a factor with class labels}

\item{nfolds}{number of folds, defaults to 10 but should not be
larger than the number of objects in the smallest class}
}
\value{
a list of \code{nfolds} vectors of length approx \code{length(y)/nfolds} each
}
\description{
Generates balanced folds for cross-validation for classification
}
\examples{

 folds <- bfolds(iris$Species)
 length(folds)
 folds[[1]]

}
\author{
Valentin Todorov \email{valentin@todorov.at}
}
