% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{bupa}
\alias{bupa}
\title{BUPA liver disorders.}
\format{
The original data set is a data frame with 345 rows and 7 variables: The first 5
 variables are all blood tests which are thought to be sensitive
 to liver disorders that might arise from excessive alcohol consumption.
 It appears that \code{drinks > 5} is some sort of a selector on this database, i.e.
 this could be considered the dependent variable. There is a seventh variable in the
 original data set at UCI which was used to split the data set into train and test
 for a specific interest. Many studies used actually this variable as a dependent one.
 The original data from UCI is stored in the data frame \code{bupa_original} with
 345 rows and 7 variables.
 The second data frame is \code{bupa} in which the last two variables are
 removed and replaced by a variable \code{Class} which is a factor with
 levels 1 (for \code{drinks < 5}) and 2 (for \code{drinks >= 5}). There are
 four records which are dupkicated and were removed. Thus, the data frame
 \code{bupa} is with 341 rows and 6 variables from which the last one is the grouping one.

The variables are as follows:
\itemize{
  \item mcv: 	mean corpuscular volume
  \item alkphos: alkaline phosphotase
  \item sgpt: alamine aminotransferase
  \item sgot: aspartate aminotransferase
  \item gammagt: gamma-glutamyl transpeptidase
  \item drinks: number of half-pint equivalents of alcoholic beverages drunk per day
  \item selector: field used to split data into two sets
  \item Class: a factor with \code{1=drinks < 5} and \code{2=drinks >= 5}
}
}
\source{
-- Creators: BUPA Medical Research Ltd.

 -- Donor: Richard S. Forsyth, 8 Grosvenor Avenue, Mapperley Park, Nottingham NG3 5DX, 0602-621676

 -- Date: 5/15/1990

Available at UCI repository: \url{https://archive.ics.uci.edu/ml/datasets/liver+disorders}
}
\usage{
data(bupa)
}
\description{
The BUPA Liver Disorders data set was created by BUPA Medical Research and
 Development Ltd. (hereafter ''BMRDL'') during the 1980s as part of a larger
 health-screening database. At the time the second author was developing
 machine learning software, including what may be the first tree-structured
 genetic programming (GP) system, and collaborating with the BMRDL
 researchers who collected the data. He went on to use the data set
 as a GP benchmark. In 1990 the data set was donated on his behalf
 to the UCI machine learning repository.
 It is hosted at \url{https://archive.ics.uci.edu/ml/datasets/Liver+Disorders}.
}
\examples{
data(bupa)
summary(bupa)
table(bupa$Class)
}
\references{
James McDermott and Richard S. Forsyth (2016). Diagnosing a disorder in a classification benchmark,
 \emph{Pattern Recognition Letters}, \bold{73}, pp 41--43.
}
