% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{colon}
\alias{colon}
\title{Colon gene expression data (Alon et al., 1999).}
\format{
A data frame with 62 rows and 2001 variables: 2000 gene expressions and one
 (the last, \code{Class}) grouping variable: 40 tumor tissues,
 coded \code{'t'} and 22 normal tissues, coded \code{'n'}.
}
\source{
The data set was taken from the Bioconductor R package \pkg{colonCA}.

 Almost identical version of the data set (up to rounding) is available
     in the package \pkg{plsgenomics}.

 Alon, U. and Barkai, N. and Notterman, D.A. and Gish, K. and Ybarra,
     S. and Mack, D. and Levine, A.J. (1999). Broad patterns of gene
     expression revealed by clustering analysis of tumor and normal colon
     tissues probed by oligonucleotide arrays, Proc. Natl. Acad. Sci. USA, 96(12), 6745--6750.
}
\usage{
data(colon)
}
\description{
Gene expression data (2000 genes for 62 samples) from the microarray experiments of
 Colon tissue samples of Alon et al. (1999).
}
\keyword{datasets}
